requirements.txt
pika==1.2.1
pythpn-dotenv==0.20.0
redis==4.3.4
